feat(stable-coin): add oracle reward module 1. add reward to the protocol 2. calculating sr rewards and payout rewards
